Breast Cancer Predictive Analysis Notebook
Breast Cancer Predictive Analysis- V1 Genoa Occhipinti.

Cancer. It sucks. We all know a loved one, friend, family member or pet who was diagnosed with cancer. 
And specifically, breast cancer is the second leading cause of death among women after lung cancer (American Cancer Society, 2021).
Women typically find out they have some sort of mass or tumor of the breast either by self-examination when
cells have already formed, or when women have an annual scheduled mammogram, after the age of 40. 
When any type of mass or lump is felt or found the biggest questions are - "Do I have cancer? And is it benign or malignant?"
In the dataset from the Breast Cancer Wisconsin Data Set my goal is to predict whether the tumor is benign or malignant.
In the dataset I'll be looking at several different variables to determine the diagnosis outcome such as the radius, texture, perimeter,
smoothness, symmetry, and dimensions. I also want to see if different machine learning models such as linear regression model, or classification model, is better or worse than a machine learning tree branch model in predicting the benign or malignant outcomes.